Extract ExternalServer logic out of junit Extension
Just moving the code into a different class
